requirements.txt
dnspython>=2.4.2
python-whois>=0.8.0
requests>=2.31.0
jinja2>=3.1.2
geoip2>=4.7.0
qrcode>=7.4.2
pillow>=10.0.0
rich>=13.5.2
